+++ cmd/shortener/main.go
@@ -5,19 +5,19 @@ import (
 	"log"
 	_ "net/http/pprof"
 
+	"github.com/MagicNetLab/ya-practicum-shortener/internal/app/repo"
 	"github.com/MagicNetLab/ya-practicum-shortener/internal/app/server"
-	"github.com/MagicNetLab/ya-practicum-shortener/internal/app/storage"
 	"github.com/MagicNetLab/ya-practicum-shortener/internal/config"
 	"github.com/MagicNetLab/ya-practicum-shortener/internal/service/logger"
 )
 
 var (
-	buildVersion string = "N/A"
-	buildDate    string = "N/A"
-	buildCommit  string = "N/A"
+	buildVersion = "N/A"
+	buildDate    = "N/A"
+	buildCommit  = "N/A"
 )
 
-func main() {
+func init() {
 	fmt.Printf("Build version: %s\n", buildVersion)
 	fmt.Printf("Build date: %s\n", buildDate)
 	fmt.Printf("Build commit: %s\n", buildCommit)
@@ -25,20 +25,23 @@ func main() {
 	err := logger.Initialize()
 	if err != nil {
 		log.Fatalf("Failed to initialize logger: %v", err)
-	}
-	defer logger.Sync()
-
-	conf := config.GetParams()
-	if !conf.IsValid() {
-		logger.Fatal("Invalid server parameters. App is not running.", nil)
 		return
 	}
 
-	_, err = storage.GetStore()
+	err = config.Initialize()
+	if err != nil {
+		args := map[string]interface{}{"error": err.Error()}
+		logger.Fatal("Failed to initialize config", args)
+	}
+
+	err = repo.Initialize(config.GetParams())
 	if err != nil {
 		args := map[string]interface{}{"error": err.Error()}
 		logger.Fatal("Failed to initialize storage", args)
 	}
 
-	server.Run(conf)
+}
+
+func main() {
+	server.Run(config.GetParams())
 }
