+++ internal/app/server/handlers/web_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/MagicNetLab/ya-practicum-shortener/internal/app/storage"
+	"github.com/MagicNetLab/ya-practicum-shortener/internal/app/repo"
 	"github.com/MagicNetLab/ya-practicum-shortener/internal/config"
 	"github.com/MagicNetLab/ya-practicum-shortener/internal/service/jwttoken"
 )
@@ -91,6 +91,11 @@ func Test_encodeLinkHeader(t *testing.T) {
 		},
 	}
 
+	err := config.Initialize()
+	assert.NoError(t, err)
+	err = repo.Initialize(config.GetParams())
+	assert.NoError(t, err)
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			request := httptest.NewRequest(tt.method, tt.request, strings.NewReader(tt.body))
@@ -124,12 +129,15 @@ func Test_encodeLinkHeader(t *testing.T) {
 
 func Test_encodeLinkByUnique(t *testing.T) {
 	t.Run("test send not unique link", func(t *testing.T) {
-		store, err := storage.GetStore()
+		err := config.Initialize()
 		assert.NoError(t, err)
+		err = repo.Initialize(config.GetParams())
+		assert.NoError(t, err)
+
 		link := "https://cloud.ru"
 		userID := 3
 
-		err = store.PutLink(context.Background(), link, "uweyiu", userID)
+		err = repo.PutLink(context.Background(), link, "uweyiu", userID)
 		assert.NoError(t, err)
 
 		request := httptest.NewRequest(http.MethodPost, "/", strings.NewReader(link))
@@ -200,16 +208,18 @@ func Test_decodeLinkHeader(t *testing.T) {
 		// todo success get link
 	}
 
+	err := config.Initialize()
+	assert.NoError(t, err)
+	err = repo.Initialize(config.GetParams())
+	assert.NoError(t, err)
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if tt.deleted {
-				store, err := storage.GetStore()
+				err := repo.PutLink(context.Background(), "http://test.link", "jsdhkahs", 3)
 				assert.NoError(t, err)
 
-				err = store.PutLink(context.Background(), "http://test.link", "jsdhkahs", 3)
-				assert.NoError(t, err)
-
-				err = store.DeleteBatchLinksArray(context.Background(), []string{"jsdhkahs"}, 3)
+				err = repo.DeleteBatchLinksArray(context.Background(), []string{"jsdhkahs"}, 3)
 				assert.NoError(t, err)
 			}
