--- internal/app/storage/local/cache.go
@@ -1,8 +1,0 @@
-package local
-
-var storeFile cacheStore
-
-// getCacheStore получение кэша с данными
-func getCacheStore() *cacheStore {
-	return &storeFile
-}
