--- internal/app/storage/local/store.go
@@ -1,4 +1,0 @@
-package local
-
-// Store локальное хранилище ссылок в памяти приложения
-var Store = store{store: make(map[string]linkEntity, 2)}
